feat: server separation
