+++ web/app/components/next_statistics/loader.server.ts
@@ -1,9 +1,8 @@
 import type { SerializeFrom } from '@remix-run/node';
 import { prismaClient } from '~/prismaClient';
-import { addDays, eachDayOfInterval, endOfDay, format, isDate, isSameDay, set, setDay, startOfDay, subMinutes } from 'date-fns';
+import { eachDayOfInterval, format, isSameDay, set, setDay, subMinutes } from 'date-fns';
 
-/** Number of minutes backwards that should count towards current visitors */
-export const CURRENT_VISITORS_LAST_MINUTES = 5;
+import { CURRENT_VISITORS_LAST_MINUTES, formatFilterDate, getFilteringParams } from './utils';
 
 export type LoadStatistics = Awaited<ReturnType<typeof loadStatistics>>;
 export type LoadStatisticsSerialized = SerializeFrom<Awaited<ReturnType<typeof loadStatistics>>>;
@@ -77,30 +76,6 @@ export const loadStatistics = async ({ request, teamSlug, projectSlug }: LoadSta
     totalPageViews,
     uniqueVisitorsTrend,
   };
-};
-
-const formatFilterDate = (date: Date) => format(date, 'yyyy-MM-dd');
-
-const getDateFromSearchParam = (param: string | null) => {
-  if (!param) return undefined;
-  const date = new Date(param);
-  if (!isDate(date)) return undefined;
-  return date;
-};
-
-const PERIOD_TYPES = ['day', 'hour'] as const;
-type PERIOD = (typeof PERIOD_TYPES)[number];
-
-const getFilteringParams = (request: Request) => {
-  const searchParams = new URL(request.url).searchParams;
-  const rawPeriod = searchParams.get('period');
-  const period: PERIOD = (rawPeriod && PERIOD_TYPES.includes(rawPeriod as PERIOD) ? rawPeriod : 'day') as PERIOD;
-
-  const defaultGte = new Date();
-  const dateGte = startOfDay(getDateFromSearchParam(searchParams.get('gte')) || addDays(defaultGte, -30));
-  const dateLte = endOfDay(getDateFromSearchParam(searchParams.get('lte')) || defaultGte);
-
-  return { dateGte, dateLte, period };
 };
 
 type WhereQuery = Awaited<ReturnType<typeof getWhereQuery>>;
