+++ web/remix.config.js
@@ -2,13 +2,12 @@
 module.exports = {
   ignoredRouteFiles: ['**/.*'],
   serverDependenciesToBundle: [
-    'd3-time-format',
-    'd3-color',
-    '@nivo/colors',
-    '@nivo/line',
-    '@nivo/bar',
-    '@nivo/core',
-    '@nivo/legends',
+    /^@nivo\/.*/,
+    /^@react-spring\/.*/,
+    /^d3-.*/,
+    'delaunator',
+    'internmap',
+    'react-lifecycles-compat',
     '@olros/stats',
     /^remix-utils.*/,
     'is-ip',
